let's see if this relief the metadata problem (bug 2845)
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/XMLSpider@24921 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/plugins/XMLSpider/IndexWriter.java
@@ -305,7 +305,8 @@ public class IndexWriter {
 				
 				if ((count > 1 && (estimateSize + pages.size() * 13) > MAX_SIZE) || //
 						(count > MAX_ENTRIES)) {
-					return false;
+					if (prefix.length() < 3 && indices.size() < 256) // FIXME this is a hack to limit number of files. remove after metadata fix
+						return false;
 				}
 
 				for (Page page : pages) {
@@ -395,8 +396,10 @@ public class IndexWriter {
 		}
 		
 		if (outputFile.length() > MAX_SIZE && count > 1) {
-			outputFile.delete();
-			return false;
+			if (prefix.length() < 3 && indices.size() < 256) { // FIXME this is a hack to limit number of files. remove after metadata fix
+				outputFile.delete();
+				return false;
+			}
 		}
 
 		if (logMINOR)
